Invalid

+++ verilator/src/memory/npc.cpp
@@ -6,6 +6,6 @@
 #include "common.h"
 #include "memory/pmem.h"
 
-void npc_mem_ifetch(const long long addr, long long *inst) {
+void npc_mem_ifetch(const long long addr, int *inst) {
   *inst = paddr_read(addr, 4);
 }
